chore: refactor to use test suite
