+++ tests/integration/liveness_test.go
@@ -3,329 +3,31 @@ package integration_test
 import (
 	"context"
 	"encoding/hex"
-	"encoding/json"
 	"fmt"
-	"github.com/celestiaorg/tastora/framework/testutil/broadcast"
-	"github.com/multiformats/go-multihash"
-
-	//"github.com/celestiaorg/tastora/framework/testutil/config"
-	"github.com/celestiaorg/tastora/framework/testutil/sdkacc"
-	"github.com/celestiaorg/tastora/framework/testutil/wallet"
-	//cometcfg "github.com/cometbft/cometbft/config"
-	"github.com/cometbft/cometbft/crypto"
-	cmtjson "github.com/cometbft/cometbft/libs/json"
-	cmprivval "github.com/cometbft/cometbft/privval"
-	"github.com/cosmos/cosmos-sdk/types/module/testutil"
-	"github.com/cosmos/cosmos-sdk/x/auth"
-	"github.com/cosmos/cosmos-sdk/x/bank"
-	"github.com/libp2p/go-libp2p/core/peer"
-	"go.uber.org/zap/zaptest"
-	"os"
 	"testing"
-	"time"
 
 	"cosmossdk.io/math"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/credentials/insecure"
-
 	"github.com/celestiaorg/go-square/v2/share"
 	"github.com/celestiaorg/tastora/framework/docker"
-	"github.com/celestiaorg/tastora/framework/docker/container"
+	"github.com/celestiaorg/tastora/framework/testutil/broadcast"
+	"github.com/celestiaorg/tastora/framework/testutil/sdkacc"
 	"github.com/celestiaorg/tastora/framework/testutil/wait"
+	"github.com/celestiaorg/tastora/framework/testutil/wallet"
 	"github.com/celestiaorg/tastora/framework/types"
-	"github.com/moby/moby/client"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
 	"github.com/stretchr/testify/require"
+	"github.com/stretchr/testify/suite"
+	"google.golang.org/grpc"
+	"google.golang.org/grpc/credentials/insecure"
 )
 
 const (
-	denom            = "stake"
-	celestiaAppTag   = "v4.0.10"
-	celestiaNodeTag  = "v0.23.5"
-	celestiaNodeRepo = "ghcr.io/celestiaorg/celestia-node"
-	celestiaAppRepo  = "ghcr.io/celestiaorg/celestia-app"
+	denom = "stake"
 )
 
-// CreateCelestiaChain sets up a Celestia app chain for DA
-func CreateCelestiaChain(ctx context.Context, t *testing.T, dockerClient *client.Client, networkID string) (*docker.Chain, error) {
-	testEncCfg := testutil.MakeTestEncodingConfig(auth.AppModuleBasic{}, bank.AppModuleBasic{})
-	celestia, err := docker.NewChainBuilder(t).
-		WithEncodingConfig(&testEncCfg).
-		WithDockerClient(dockerClient).
-		WithDockerNetworkID(networkID).
-		WithImage(container.NewImage(celestiaAppRepo, celestiaAppTag, "10001:10001")).
-		WithAdditionalStartArgs(
-			"--force-no-bbr",
-			"--grpc.enable",
-			"--grpc.address", "0.0.0.0:9090",
-			"--rpc.grpc_laddr=tcp://0.0.0.0:9098",
-			"--timeout-commit", "1s",
-			"--minimum-gas-prices", "0.000001utia",
-		).
-		WithNode(docker.NewChainNodeConfigBuilder().Build()).
-		Build(ctx)
-
-	if err != nil {
-		return nil, fmt.Errorf("failed to build celestia chain: %w", err)
-	}
-
-	err = celestia.Start(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("failed to start celestia chain: %w", err)
-	}
-
-	return celestia, nil
-}
-
-// CreateDANetwork sets up the DA network with bridge and full nodes
-func CreateDANetwork(ctx context.Context, t *testing.T, dockerClient *client.Client, networkID string, celestiaChain *docker.Chain) (types.DataAvailabilityNetwork, types.DANode, error) {
-	config := &docker.Config{
-		Logger:          zaptest.NewLogger(t),
-		DockerClient:    dockerClient,
-		DockerNetworkID: networkID,
-		DataAvailabilityNetworkConfig: &docker.DataAvailabilityNetworkConfig{
-			Image:           container.NewImage(celestiaNodeRepo, celestiaNodeTag, "10001:10001"),
-			BridgeNodeCount: 1,
-		},
-	}
-
-	provider := docker.NewProvider(*config, t)
-	daNetwork, err := provider.GetDataAvailabilityNetwork(ctx)
-	if err != nil {
-		return nil, nil, fmt.Errorf("failed to get DA network: %w", err)
-	}
-
-	genesisHash, err := getGenesisHash(ctx, celestiaChain)
-	if err != nil {
-		return nil, nil, fmt.Errorf("failed to get genesis hash: %w", err)
-	}
-
-	bridgeNodes := daNetwork.GetBridgeNodes()
-	if len(bridgeNodes) == 0 {
-		return nil, nil, fmt.Errorf("no bridge nodes available")
-	}
-
-	bridgeNode := bridgeNodes[0]
-
-	chainID := celestiaChain.GetChainID()
-	celestiaNodeHostname, err := celestiaChain.GetNodes()[0].GetInternalHostName(ctx)
-	if err != nil {
-		return nil, nil, fmt.Errorf("failed to get celestia node hostname: %w", err)
-	}
-
-	err = bridgeNode.Start(ctx,
-		types.WithChainID(chainID),
-		// my-key is the name of the keyring generated by the bridge node, we can fund this key so that we can submit blobs.
-		types.WithAdditionalStartArguments("--p2p.network", chainID, "--core.ip", celestiaNodeHostname, "--rpc.addr", "0.0.0.0", "--keyring.keyname", "my-key"),
-		types.WithEnvironmentVariables(map[string]string{
-			"CELESTIA_CUSTOM": types.BuildCelestiaCustomEnvVar(chainID, genesisHash, ""),
-			"P2P_NETWORK":     chainID,
-		}),
-	)
-	if err != nil {
-		return nil, nil, fmt.Errorf("failed to start bridge node: %w", err)
-	}
-
-	// hack to get around global, need to set the address prefix before use.
-	sdk.GetConfig().SetBech32PrefixForAccount("celestia", "celestiapub")
-
-	// Fund the bridge node DA wallet to enable blob submission
-	t.Log("Funding bridge node DA wallet...")
-	fundingWallet := celestiaChain.GetFaucetWallet()
-
-	// Get the bridge node's wallet
-	bridgeWallet, err := bridgeNode.GetWallet()
-	require.NoError(t, err, "failed to get bridge node wallet")
-
-	// fund the bridge node wallet
-	daFundingAmount := sdk.NewCoins(sdk.NewCoin("utia", math.NewInt(10_000_000)))
-	err = sendFunds(ctx, celestiaChain, fundingWallet, bridgeWallet, daFundingAmount, 0)
-	require.NoError(t, err, "failed to fund bridge node DA wallet")
-
-	return daNetwork, bridgeNode, nil
-}
-
-// CreateRollkitChain sets up the rollkit chain connected to the DA network
-func CreateRollkitChain(ctx context.Context, t *testing.T, dockerClient *client.Client, networkID string, bridgeNode types.DANode, celestiaChain types.Chain) (*docker.Chain, error) {
-	// Get DA connection details
-	authToken, err := bridgeNode.GetAuthToken()
-	if err != nil {
-		return nil, fmt.Errorf("failed to get auth token: %w", err)
-	}
-
-	bridgeRPCAddress, err := bridgeNode.GetInternalRPCAddress()
-	if err != nil {
-		return nil, fmt.Errorf("failed to get bridge RPC address: %w", err)
-	}
-
-	daAddress := fmt.Sprintf("http://%s", bridgeRPCAddress)
-	namespace := generateValidNamespace()
-
-	celestiaHeight, err := celestiaChain.Height(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get celestia chain height: %w", err)
-	}
-	daStartHeight := fmt.Sprintf("%d", celestiaHeight)
-
-	// create the single sequencer genesis function for the aggregator node
-	addSingleSequencer := CreateAddSingleSequencer(daAddress, authToken, namespace, daStartHeight)
-
-	// bank and auth modules required to deal with bank send tx's
-	testEncCfg := testutil.MakeTestEncodingConfig(auth.AppModuleBasic{}, bank.AppModuleBasic{})
-	// Create chain with only the aggregator node initially
-	rollkitChain, err := docker.NewChainBuilder(t).
-		WithEncodingConfig(&testEncCfg).
-		WithImage(getRollkitAppContainer()).
-		WithDenom(denom).
-		WithDockerClient(dockerClient).
-		WithName("rollkit").
-		WithDockerNetworkID(networkID).
-		WithChainID("rollkit-test").
-		WithBech32Prefix("gm").
-		WithBinaryName("gmd").
-		// explicitly set 0 gas so that we can make exact assertions when sending balances.
-		WithGasPrices(fmt.Sprintf("0.00%s", denom)).
-		WithNodes(
-			docker.NewChainNodeConfigBuilder().
-				// Create aggregator node with rollkit-specific start arguments
-				WithAdditionalStartArgs(
-					"--rollkit.node.aggregator",
-					"--rollkit.signer.passphrase", "12345678",
-					"--rollkit.da.address", daAddress,
-					"--rollkit.da.gas_price", "0.000001",
-					"--rollkit.da.auth_token", authToken,
-					"--rollkit.rpc.address", "0.0.0.0:7331",
-					"--rollkit.da.namespace", namespace,
-					"--rollkit.da.start_height", daStartHeight,
-					"--rollkit.p2p.listen_address", "/ip4/0.0.0.0/tcp/36656",
-					"--log_level", "*:debug,rollkit:debug,p2p:debug",
-				).
-				WithGenTX(true).
-				WithPostInit(addSingleSequencer).
-				Build(),
-		).
-		Build(ctx)
-
-	if err != nil {
-		return nil, fmt.Errorf("failed to build rollkit chain: %w", err)
-	}
-
-	// Start ONLY the aggregator node first and let it stabilize
-	err = rollkitChain.Start(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("failed to start rollkit chain: %w", err)
-	}
-
-	// Wait for aggregator to fully start and be ready
-	t.Logf("Waiting for aggregator node to stabilize...")
-	time.Sleep(10 * time.Second)
-
-	// Query the aggregator node ID for P2P connectivity
-	aggregatorNode := rollkitChain.Nodes()[0] // aggregator is the first (and only) node
-	aggregatorNodeID, err := aggregatorNode.NodeID(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get aggregator node ID: %w", err)
-	}
-	require.NotEmpty(t, aggregatorNodeID, "aggregator node ID is empty")
-
-	// Get the aggregator's internal IP address
-	aggregatorIP, err := aggregatorNode.GetInternalIP(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get aggregator internal IP: %w", err)
-	}
-
-	// Convert hex node ID to libp2p peer ID
-	nodeIDBytes, err := hex.DecodeString(aggregatorNodeID)
-	if err != nil {
-		return nil, fmt.Errorf("failed to decode node ID: %w", err)
-	}
-
-	// Create a multihash from the node ID bytes (using identity hash since it's already hashed)
-	mh, err := multihash.Sum(nodeIDBytes, multihash.IDENTITY, -1)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create multihash: %w", err)
-	}
-
-	// Create peer ID from multihash
-	peerID, err := peer.IDFromBytes(mh)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create peer ID: %w", err)
-	}
-
-	// Construct the aggregator peer in multiaddr format
-	aggregatorPeer := fmt.Sprintf("/ip4/%s/tcp/36656/p2p/%s", aggregatorIP, peerID.String())
-	t.Logf("Constructed aggregator peer: %s", aggregatorPeer)
-	t.Logf("Aggregator Node ID (hex): %s", aggregatorNodeID)
-	t.Logf("Aggregator Node ID (base58): %s", peerID.String())
-	t.Logf("Aggregator IP: %s", aggregatorIP)
-	t.Logf("Aggregator Hostname: %s", aggregatorNode.HostName())
-
-	// Now add follower nodes one by one, with delays between them
-	t.Logf("Adding first follower node...")
-	err = rollkitChain.AddNode(ctx, docker.NewChainNodeConfigBuilder().
-		WithAdditionalStartArgs(
-			"--rollkit.da.address", daAddress,
-			"--rollkit.da.gas_price", "0.000001",
-			"--rollkit.da.auth_token", authToken,
-			"--rollkit.rpc.address", "0.0.0.0:7331",
-			"--rollkit.da.namespace", namespace,
-			"--rollkit.da.start_height", daStartHeight,
-			"--rollkit.p2p.listen_address", "/ip4/0.0.0.0/tcp/36656",
-			//"--rollkit.p2p.peers", aggregatorPeer,
-			"--log_level", "*:debug,rollkit:debug,p2p:debug",
-		).
-		WithGenTX(false).
-		Build())
-	if err != nil {
-		return nil, fmt.Errorf("failed to add first follower node: %w", err)
-	}
-
-	// Wait a bit before adding the second follower
-	time.Sleep(5 * time.Second)
-
-	t.Logf("Adding second follower node...")
-	err = rollkitChain.AddNode(ctx, docker.NewChainNodeConfigBuilder().
-		WithAdditionalStartArgs(
-			"--rollkit.da.address", daAddress,
-			"--rollkit.da.gas_price", "0.000001",
-			"--rollkit.da.auth_token", authToken,
-			"--rollkit.rpc.address", "0.0.0.0:7331",
-			"--rollkit.da.namespace", namespace,
-			"--rollkit.da.start_height", daStartHeight,
-			"--rollkit.p2p.listen_address", "/ip4/0.0.0.0/tcp/36656",
-			//"--rollkit.p2p.peers", aggregatorPeer,
-			"--log_level", "*:debug,rollkit:debug,p2p:debug",
-		).
-		WithGenTX(false).
-		Build())
-	if err != nil {
-		return nil, fmt.Errorf("failed to add second follower node: %w", err)
-	}
-
-	return rollkitChain, nil
-}
-
-// getGenesisHash retrieves the genesis hash from the celestia chain
-func getGenesisHash(ctx context.Context, celestiaChain types.Chain) (string, error) {
-	node := celestiaChain.GetNodes()[0]
-	c, err := node.GetRPCClient()
-	if err != nil {
-		return "", fmt.Errorf("failed to get node client: %w", err)
-	}
-
-	first := int64(1)
-	block, err := c.Block(ctx, &first)
-	if err != nil {
-		return "", fmt.Errorf("failed to get block: %w", err)
-	}
-
-	genesisHash := block.Block.Header.Hash().String()
-	if genesisHash == "" {
-		return "", fmt.Errorf("genesis hash is empty")
-	}
-
-	return genesisHash, nil
+func TestDockerIntegrationTestSuite(t *testing.T) {
+	suite.Run(t, new(DockerIntegrationTestSuite))
 }
 
 func generateValidNamespace() string {
@@ -382,6 +84,20 @@ func sendFunds(ctx context.Context, chain *docker.Chain, fromWallet, toWallet ty
 	return nil
 }
 
+// TestLivenessWithCelestiaDA tests the liveness of rollkit with Celestia DA
+func (suite *DockerIntegrationTestSuite) TestLivenessWithCelestiaDA() {
+	ctx := context.Background()
+	// Test block production - wait for rollkit chain to produce blocks
+	suite.T().Log("Testing block production...")
+	suite.Require().NoError(wait.ForBlocks(ctx, 5, suite.rollkitChain))
+
+	// Test transaction submission and query
+	suite.T().Log("Testing transaction submission and query...")
+	testTransactionSubmissionAndQuery(suite.T(), ctx, suite.rollkitChain)
+
+	suite.T().Log("Test completed successfully")
+}
+
 // testTransactionSubmissionAndQuery tests sending transactions and querying results using tastora API
 func testTransactionSubmissionAndQuery(t *testing.T, ctx context.Context, rollkitChain *docker.Chain) {
 	// hack to get around global, need to set the address prefix before use.
@@ -414,178 +130,4 @@ func testTransactionSubmissionAndQuery(t *testing.T, ctx context.Context, rollki
 	carolBalance, err := queryBankBalance(ctx, rollkitChain.GetGRPCAddress(), carolsWallet.GetFormattedAddress(), denom)
 	require.NoError(t, err, "failed to query carol's balance")
 	require.True(t, carolBalance.Amount.Equal(math.NewInt(100)), "carol should have received 100 tokens")
-}
-
-func TestLivenessWithCelestiaDA(t *testing.T) {
-	ctx := context.Background()
-	dockerClient, networkID := docker.DockerSetup(t)
-
-	celestiaChain, err := CreateCelestiaChain(ctx, t, dockerClient, networkID)
-	require.NoError(t, err)
-	t.Cleanup(func() {
-		if err := celestiaChain.Stop(ctx); err != nil {
-			t.Logf("failed to stop celestia chain: %v", err)
-		}
-	})
-
-	t.Log("Celestia app chain started")
-
-	_, bridgeNode, err := CreateDANetwork(ctx, t, dockerClient, networkID, celestiaChain)
-	require.NoError(t, err)
-
-	t.Log("Bridge node started")
-
-	rollkitChain, err := CreateRollkitChain(ctx, t, dockerClient, networkID, bridgeNode, celestiaChain)
-	require.NoError(t, err)
-	t.Cleanup(func() {
-		if err := rollkitChain.Stop(ctx); err != nil {
-			t.Logf("failed to stop rollkit chain: %v", err)
-		}
-	})
-
-	t.Log("Rollkit chain started")
-
-	// Debug: Check node states before testing
-	t.Log("Debug: Checking rollkit chain node states...")
-	for i, node := range rollkitChain.Nodes() {
-		hostname, _ := node.GetInternalHostName(ctx)
-		ip, _ := node.GetInternalIP(ctx)
-		t.Logf("Node %d - Hostname: %s, IP: %s", i, hostname, ip)
-	}
-
-	// Add a small delay to let P2P connections stabilize
-	t.Log("Waiting for P2P connections to stabilize...")
-	time.Sleep(5 * time.Second)
-
-	// Test block production - wait for rollkit chain to produce blocks
-	t.Log("Testing block production...")
-	require.NoError(t, wait.ForBlocks(ctx, 5, rollkitChain))
-
-	// Test transaction submission and query
-	t.Log("Testing transaction submission and query...")
-	testTransactionSubmissionAndQuery(t, ctx, rollkitChain)
-
-	// Debug: Test with just aggregator first
-	t.Log("Testing aggregator alone for 30 seconds...")
-	time.Sleep(30 * time.Second)
-
-	// Check if we can query the aggregator's net_info
-	height, _ := rollkitChain.Height(ctx)
-	t.Logf("Current chain height: %d", height)
-
-	time.Sleep(time.Hour)
-}
-
-// getPubKey returns the validator public key.
-func getPubKey(ctx context.Context, chainNode *docker.ChainNode) (crypto.PubKey, error) {
-	keyJSONBytes, err := chainNode.ReadFile(ctx, "config/priv_validator_key.json")
-	if err != nil {
-		return nil, err
-	}
-	var pvKey cmprivval.FilePVKey
-	if err = cmtjson.Unmarshal(keyJSONBytes, &pvKey); err != nil {
-		return nil, fmt.Errorf("failed to unmarshal priv_validator_key.json: %w", err)
-	}
-	return pvKey.PubKey, nil
-}
-
-// getRollkitAppContainer returns the rollkit app container image.
-// uses the ROLLKIT_IMAGE_REPO and ROLLKIT_IMAGE_TAG environment variables.
-func getRollkitAppContainer() container.Image {
-	// get image repo and tag from environment variables
-	imageRepo := os.Getenv("ROLLKIT_IMAGE_REPO")
-	if imageRepo == "" {
-		imageRepo = "rollkit-gm" // fallback default
-	}
-
-	imageTag := os.Getenv("ROLLKIT_IMAGE_TAG")
-	if imageTag == "" {
-		imageTag = "latest" // fallback default
-	}
-	return container.NewImage(imageRepo, imageTag, "10001:10001")
-}
-
-// CreateAddSingleSequencer returns a closure that stores the first node's genesis configuration
-// and applies it to all subsequent nodes to ensure they have identical validator setups.
-func CreateAddSingleSequencer(daAddress, authToken, namespace, daStartHeight string) func(context.Context, *docker.ChainNode) error {
-	var storedAggregatorGenesis []byte
-
-	return func(ctx context.Context, node *docker.ChainNode) error {
-
-		// if we already have stored genesis from the aggregator, use it for followers
-		if storedAggregatorGenesis != nil {
-			err := node.WriteFile(ctx, "config/genesis.json", storedAggregatorGenesis)
-			if err != nil {
-				return fmt.Errorf("failed to write genesis.json: %w", err)
-			}
-			return nil
-		}
-
-		// this is the first call (aggregator node), process and store the genesis
-		return addSingleSequencerFirstTime(ctx, node, &storedAggregatorGenesis, daAddress, authToken, namespace, daStartHeight)
-	}
-}
-
-// addSingleSequencerFirstTime modifies the genesis file to add a single sequencer with specified power and public key.
-// Reads the genesis file from the node, updates the validators with the sequencer info, and writes the updated file back.
-// Also removes any additional validators from genesis transactions to ensure only one validator exists.
-func addSingleSequencerFirstTime(ctx context.Context, node *docker.ChainNode, storedGenesis *[]byte, daAddress, authToken, namespace, daStartHeight string) error {
-	genesisBz, err := node.ReadFile(ctx, "config/genesis.json")
-	if err != nil {
-		return fmt.Errorf("failed to read genesis.json: %w", err)
-	}
-
-	pubKey, err := getPubKey(ctx, node)
-	if err != nil {
-		return fmt.Errorf("failed to get pubkey: %w", err)
-	}
-
-	var genDoc map[string]interface{}
-	if err := json.Unmarshal(genesisBz, &genDoc); err != nil {
-		return fmt.Errorf("failed to parse genesis.json: %w", err)
-	}
-
-	// remove additional validators from genesis transactions
-	if appState, ok := genDoc["app_state"].(map[string]interface{}); ok {
-		if genutil, ok := appState["genutil"].(map[string]interface{}); ok {
-			if genTxs, ok := genutil["gen_txs"].([]interface{}); ok && len(genTxs) > 1 {
-				// keep only the first genesis transaction to ensure single validator
-				genutil["gen_txs"] = []interface{}{genTxs[0]}
-			}
-		}
-	}
-
-	// set consensus validators to only include the first validator
-	consensus, ok := genDoc["consensus"].(map[string]interface{})
-	if !ok {
-		return fmt.Errorf("genesis.json does not contain a valid 'consensus' object")
-	}
-	consensus["validators"] = []map[string]interface{}{
-		{
-			"name":    "Rollkit Sequencer",
-			"address": pubKey.Address(),
-			"pub_key": map[string]interface{}{
-				"type":  "tendermint/PubKeyEd25519",
-				"value": pubKey.Bytes(),
-			},
-			"power": "5", // NOTE: because of default validator wallet amount in tastora the power will be computed as 5.
-		},
-	}
-
-	updatedGenesis, err := json.MarshalIndent(genDoc, "", "  ")
-	if err != nil {
-		return fmt.Errorf("failed to marshal genesis: %w", err)
-	}
-
-	// store the genesis for subsequent nodes
-	*storedGenesis = updatedGenesis
-
-	err = node.WriteFile(ctx, "config/genesis.json", updatedGenesis)
-	if err != nil {
-		return fmt.Errorf("failed to write genesis.json: %w", err)
-	}
-
-	return nil
-	// create rollkit config for the aggregator (first node)
-	//return createRollkitConfig(ctx, node, true, daAddress, authToken, namespace, daStartHeight)
 }
