+++ tests/integration/testsuite_test.go
@@ -1,0 +1,431 @@
+package integration_test
+
+import (
+	"context"
+	"encoding/hex"
+	"encoding/json"
+	"fmt"
+	"github.com/libp2p/go-libp2p/core/peer"
+	"github.com/multiformats/go-multihash"
+	"os"
+
+	"cosmossdk.io/math"
+	"github.com/celestiaorg/tastora/framework/docker"
+	"github.com/celestiaorg/tastora/framework/docker/container"
+	"github.com/celestiaorg/tastora/framework/testutil/sdkacc"
+	"github.com/celestiaorg/tastora/framework/types"
+	"github.com/cometbft/cometbft/crypto"
+	cmtjson "github.com/cometbft/cometbft/libs/json"
+	cmprivval "github.com/cometbft/cometbft/privval"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/types/module/testutil"
+	"github.com/cosmos/cosmos-sdk/x/auth"
+	"github.com/cosmos/cosmos-sdk/x/bank"
+	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
+	"github.com/moby/moby/client"
+	"github.com/stretchr/testify/suite"
+	"go.uber.org/zap/zaptest"
+)
+
+const (
+	celestiaAppTag   = "v4.0.10"
+	celestiaNodeTag  = "v0.23.5"
+	celestiaAppRepo  = "ghcr.io/celestiaorg/celestia-app"
+	celestiaNodeRepo = "ghcr.io/celestiaorg/celestia-node"
+)
+
+// DockerIntegrationTestSuite provides common functionality for integration tests using Docker
+type DockerIntegrationTestSuite struct {
+	suite.Suite
+
+	dockerClient  *client.Client
+	networkID     string
+	celestiaChain *docker.Chain
+	daNetwork     types.DataAvailabilityNetwork
+	rollkitChain  *docker.Chain
+	bridgeNode    types.DANode
+}
+
+// SetupTest initializes the Docker environment for each test
+func (suite *DockerIntegrationTestSuite) SetupTest() {
+	ctx := context.Background()
+	suite.dockerClient, suite.networkID = docker.DockerSetup(suite.T())
+
+	suite.celestiaChain = suite.CreateCelestiaChain(ctx)
+	suite.T().Log("Celestia app chain started")
+
+	suite.bridgeNode = suite.CreateDANetwork(ctx)
+	suite.T().Log("Bridge node started")
+
+	suite.rollkitChain = suite.CreateRollkitChain(ctx)
+	suite.T().Log("Rollkit chain started")
+}
+
+// TearDownTest cleans up resources after each test
+func (suite *DockerIntegrationTestSuite) TearDownTest() {
+	ctx := context.Background()
+
+	if suite.rollkitChain != nil {
+		if err := suite.rollkitChain.Stop(ctx); err != nil {
+			suite.T().Logf("failed to stop rollkit chain: %v", err)
+		}
+		suite.rollkitChain = nil
+	}
+
+	if suite.celestiaChain != nil {
+		if err := suite.celestiaChain.Stop(ctx); err != nil {
+			suite.T().Logf("failed to stop celestia chain: %v", err)
+		}
+		suite.celestiaChain = nil
+	}
+
+	// TODO: cleanup daNetwork
+	suite.daNetwork = nil
+}
+
+// CreateCelestiaChain sets up a Celestia app chain for DA and stores it in the suite
+func (suite *DockerIntegrationTestSuite) CreateCelestiaChain(ctx context.Context) *docker.Chain {
+	testEncCfg := testutil.MakeTestEncodingConfig(auth.AppModuleBasic{}, bank.AppModuleBasic{})
+	celestia, err := docker.NewChainBuilder(suite.T()).
+		WithEncodingConfig(&testEncCfg).
+		WithDockerClient(suite.dockerClient).
+		WithDockerNetworkID(suite.networkID).
+		WithImage(container.NewImage(celestiaAppRepo, celestiaAppTag, "10001:10001")).
+		WithAdditionalStartArgs(
+			"--force-no-bbr",
+			"--grpc.enable",
+			"--grpc.address", "0.0.0.0:9090",
+			"--rpc.grpc_laddr=tcp://0.0.0.0:9098",
+			"--timeout-commit", "1s",
+			"--minimum-gas-prices", "0.000001utia",
+		).
+		WithNode(docker.NewChainNodeConfigBuilder().Build()).
+		Build(ctx)
+
+	suite.Require().NoError(err)
+
+	err = celestia.Start(ctx)
+	suite.Require().NoError(err)
+	return celestia
+}
+
+// CreateDANetwork sets up the DA network with bridge and full nodes and stores it in the suite
+func (suite *DockerIntegrationTestSuite) CreateDANetwork(ctx context.Context) types.DANode {
+	config := &docker.Config{
+		Logger:          zaptest.NewLogger(suite.T()),
+		DockerClient:    suite.dockerClient,
+		DockerNetworkID: suite.networkID,
+		DataAvailabilityNetworkConfig: &docker.DataAvailabilityNetworkConfig{
+			Image:           container.NewImage(celestiaNodeRepo, celestiaNodeTag, "10001:10001"),
+			BridgeNodeCount: 1,
+		},
+	}
+
+	provider := docker.NewProvider(*config, suite.T())
+	daNetwork, err := provider.GetDataAvailabilityNetwork(ctx)
+	suite.Require().NoError(err)
+
+	suite.daNetwork = daNetwork
+
+	genesisHash, err := getGenesisHash(ctx, suite.celestiaChain)
+	suite.Require().NoError(err)
+
+	bridgeNodes := daNetwork.GetBridgeNodes()
+	suite.Require().NotEmpty(bridgeNodes, "no bridge nodes available")
+
+	bridgeNode := bridgeNodes[0]
+
+	chainID := suite.celestiaChain.GetChainID()
+	celestiaNodeHostname, err := suite.celestiaChain.GetNodes()[0].GetInternalHostName(ctx)
+	suite.Require().NoError(err)
+
+	err = bridgeNode.Start(ctx,
+		types.WithChainID(chainID),
+		// my-key is the name of the keyring generated by the bridge node, we can fund this key so that we can submit blobs.
+		types.WithAdditionalStartArguments("--p2p.network", chainID, "--core.ip", celestiaNodeHostname, "--rpc.addr", "0.0.0.0", "--keyring.keyname", "my-key"),
+		types.WithEnvironmentVariables(map[string]string{
+			"CELESTIA_CUSTOM": types.BuildCelestiaCustomEnvVar(chainID, genesisHash, ""),
+			"P2P_NETWORK":     chainID,
+		}),
+	)
+	suite.Require().NoError(err)
+
+	// Fund the bridge node DA wallet to enable blob submission
+	suite.fundBridgeNodeWallet(ctx, bridgeNode)
+
+	return bridgeNode
+}
+
+// CreateRollkitChain sets up the rollkit chain connected to the DA network and returns it
+func (suite *DockerIntegrationTestSuite) CreateRollkitChain(ctx context.Context) *docker.Chain {
+	// Get DA connection details
+	authToken, err := suite.bridgeNode.GetAuthToken()
+	suite.Require().NoError(err)
+
+	bridgeRPCAddress, err := suite.bridgeNode.GetInternalRPCAddress()
+	suite.Require().NoError(err)
+
+	daAddress := fmt.Sprintf("http://%s", bridgeRPCAddress)
+	namespace := generateValidNamespace()
+
+	celestiaHeight, err := suite.celestiaChain.Height(ctx)
+	suite.Require().NoError(err)
+	daStartHeight := fmt.Sprintf("%d", celestiaHeight)
+
+	// use the single sequencer genesis function for the aggregator node
+
+	// bank and auth modules required to deal with bank send tx's
+	testEncCfg := testutil.MakeTestEncodingConfig(auth.AppModuleBasic{}, bank.AppModuleBasic{})
+	// Create chain with only the aggregator node initially
+	rollkitChain, err := docker.NewChainBuilder(suite.T()).
+		WithEncodingConfig(&testEncCfg).
+		WithImage(getRollkitAppContainer()).
+		WithDenom("stake").
+		WithDockerClient(suite.dockerClient).
+		WithName("rollkit").
+		WithDockerNetworkID(suite.networkID).
+		WithChainID("rollkit-test").
+		WithBech32Prefix("gm").
+		WithBinaryName("gmd").
+		// explicitly set 0 gas so that we can make exact assertions when sending balances.
+		WithGasPrices(fmt.Sprintf("0.00%s", "stake")).
+		WithNodes(
+			docker.NewChainNodeConfigBuilder().
+				// Create aggregator node with rollkit-specific start arguments
+				WithAdditionalStartArgs(
+					"--rollkit.node.aggregator",
+					"--rollkit.signer.passphrase", "12345678",
+					"--rollkit.da.address", daAddress,
+					"--rollkit.da.gas_price", "0.000001",
+					"--rollkit.da.auth_token", authToken,
+					"--rollkit.rpc.address", "0.0.0.0:7331",
+					"--rollkit.da.namespace", namespace,
+					"--rollkit.da.start_height", daStartHeight,
+					"--rollkit.p2p.listen_address", "/ip4/0.0.0.0/tcp/36656",
+					"--log_level", "*:info",
+				).
+				WithGenTX(true).
+				WithPostInit(addSingleSequencer).
+				Build(),
+		).
+		Build(ctx)
+
+	suite.Require().NoError(err)
+
+	// Start the aggregator node first so that we can query the p2p
+	err = rollkitChain.Start(ctx)
+	suite.Require().NoError(err)
+
+	aggregatorPeer := suite.GetNodeMultiAddr(ctx, rollkitChain.GetNode())
+	suite.T().Logf("Aggregator peer: %s", aggregatorPeer)
+
+	// Add follower nodes
+	suite.T().Logf("Adding first follower node...")
+	err = rollkitChain.AddNode(ctx, docker.NewChainNodeConfigBuilder().
+		WithAdditionalStartArgs(
+			"--rollkit.da.address", daAddress,
+			"--rollkit.da.gas_price", "0.000001",
+			"--rollkit.da.auth_token", authToken,
+			"--rollkit.rpc.address", "0.0.0.0:7331",
+			"--rollkit.da.namespace", namespace,
+			"--rollkit.da.start_height", daStartHeight,
+			"--rollkit.p2p.listen_address", "/ip4/0.0.0.0/tcp/36656",
+			//"--rollkit.p2p.peers", aggregatorPeer, // TODO uncomment to enable P2P, seems broken right now
+			"--log_level", "*:info",
+		).
+		WithGenTX(false).
+		Build())
+	suite.Require().NoError(err)
+
+	suite.T().Logf("Adding second follower node...")
+	err = rollkitChain.AddNode(ctx, docker.NewChainNodeConfigBuilder().
+		WithAdditionalStartArgs(
+			"--rollkit.da.address", daAddress,
+			"--rollkit.da.gas_price", "0.000001",
+			"--rollkit.da.auth_token", authToken,
+			"--rollkit.rpc.address", "0.0.0.0:7331",
+			"--rollkit.da.namespace", namespace,
+			"--rollkit.da.start_height", daStartHeight,
+			"--rollkit.p2p.listen_address", "/ip4/0.0.0.0/tcp/36656",
+			//"--rollkit.p2p.peers", aggregatorPeer, // TODO uncomment to enable P2P, seems broken right now
+			"--log_level", "*:info",
+		).
+		WithGenTX(false).
+		Build())
+	suite.Require().NoError(err)
+
+	return rollkitChain
+}
+
+// GetNodeMultiAddr extracts the multiaddr from a rollkit chain node
+// Returns the multiaddr in the format: /ip4/{IP}/tcp/36656/p2p/{PEER_ID}
+func (suite *DockerIntegrationTestSuite) GetNodeMultiAddr(ctx context.Context, node *docker.ChainNode) string {
+	// Get the node ID
+	nodeID, err := node.NodeID(ctx)
+	suite.Require().NoError(err)
+	suite.Require().NotEmpty(nodeID, "node ID is empty")
+
+	// Get the node's internal IP address
+	nodeIP, err := node.GetInternalIP(ctx)
+	suite.Require().NoError(err)
+
+	// Convert hex node ID to libp2p peer ID
+	nodeIDBytes, err := hex.DecodeString(nodeID)
+	suite.Require().NoError(err)
+
+	// Create a multihash from the node ID bytes (using identity hash since it's already hashed)
+	mh, err := multihash.Sum(nodeIDBytes, multihash.IDENTITY, -1)
+	suite.Require().NoError(err)
+
+	// Create peer ID from multihash
+	peerID, err := peer.IDFromBytes(mh)
+	suite.Require().NoError(err)
+
+	// Construct the multiaddr with default rollkit P2P port
+	multiAddr := fmt.Sprintf("/ip4/%s/tcp/36656/p2p/%s", nodeIP, peerID.String())
+
+	return multiAddr
+}
+
+// getGenesisHash retrieves the genesis hash from the celestia chain
+func getGenesisHash(ctx context.Context, chain types.Chain) (string, error) {
+	node := chain.GetNodes()[0]
+	c, err := node.GetRPCClient()
+	if err != nil {
+		return "", fmt.Errorf("failed to get node client: %w", err)
+	}
+
+	first := int64(1)
+	block, err := c.Block(ctx, &first)
+	if err != nil {
+		return "", fmt.Errorf("failed to get block: %w", err)
+	}
+
+	genesisHash := block.Block.Header.Hash().String()
+	if genesisHash == "" {
+		return "", fmt.Errorf("genesis hash is empty")
+	}
+
+	return genesisHash, nil
+}
+
+// sendFunds sends funds from one wallet to another using bank transfer
+func (suite *DockerIntegrationTestSuite) sendFunds(ctx context.Context, chain *docker.Chain, fromWallet, toWallet types.Wallet, amount sdk.Coins, nodeIdx int) error {
+	fromAddress, err := sdkacc.AddressFromWallet(fromWallet)
+	if err != nil {
+		return fmt.Errorf("failed to get sender address: %w", err)
+	}
+
+	toAddress, err := sdkacc.AddressFromWallet(toWallet)
+	if err != nil {
+		return fmt.Errorf("failed to get destination address: %w", err)
+	}
+
+	chainNode := chain.GetNodes()[nodeIdx]
+	broadcaster := docker.NewBroadcasterForNode(chain, chainNode.(*docker.ChainNode))
+
+	msg := banktypes.NewMsgSend(fromAddress, toAddress, amount)
+	resp, err := broadcaster.BroadcastMessages(ctx, fromWallet, msg)
+	if err != nil {
+		return fmt.Errorf("failed to broadcast transaction: %w", err)
+	}
+
+	if resp.Code != 0 {
+		return fmt.Errorf("transaction failed with code %d: %s", resp.Code, resp.RawLog)
+	}
+
+	return nil
+}
+
+// getRollkitAppContainer returns the rollkit app container image
+func getRollkitAppContainer() container.Image {
+	// get image repo and tag from environment variables
+	imageRepo := os.Getenv("ROLLKIT_IMAGE_REPO")
+	if imageRepo == "" {
+		imageRepo = "rollkit-gm" // fallback default
+	}
+
+	imageTag := os.Getenv("ROLLKIT_IMAGE_TAG")
+	if imageTag == "" {
+		imageTag = "latest" // fallback default
+	}
+	return container.NewImage(imageRepo, imageTag, "10001:10001")
+}
+
+// addSingleSequencer modifies the genesis file to ensure single sequencer setup
+func addSingleSequencer(ctx context.Context, node *docker.ChainNode) error {
+	genesisBz, err := node.ReadFile(ctx, "config/genesis.json")
+	if err != nil {
+		return fmt.Errorf("failed to read genesis.json: %w", err)
+	}
+
+	pubKey, err := getPubKey(ctx, node)
+	if err != nil {
+		return fmt.Errorf("failed to get pubkey: %w", err)
+	}
+
+	var genDoc map[string]interface{}
+	if err := json.Unmarshal(genesisBz, &genDoc); err != nil {
+		return fmt.Errorf("failed to parse genesis.json: %w", err)
+	}
+
+	// set consensus validators to only include the first validator
+	consensus, ok := genDoc["consensus"].(map[string]interface{})
+	if !ok {
+		return fmt.Errorf("genesis.json does not contain a valid 'consensus' object")
+	}
+	consensus["validators"] = []map[string]interface{}{
+		{
+			"name":    "Rollkit Sequencer",
+			"address": pubKey.Address(),
+			"pub_key": map[string]interface{}{
+				"type":  "tendermint/PubKeyEd25519",
+				"value": pubKey.Bytes(),
+			},
+			"power": "5",
+		},
+	}
+
+	updatedGenesis, err := json.MarshalIndent(genDoc, "", "  ")
+	if err != nil {
+		return fmt.Errorf("failed to marshal genesis: %w", err)
+	}
+
+	err = node.WriteFile(ctx, "config/genesis.json", updatedGenesis)
+	if err != nil {
+		return fmt.Errorf("failed to write genesis.json: %w", err)
+	}
+
+	return nil
+}
+
+// getPubKey returns the validator public key
+func getPubKey(ctx context.Context, chainNode *docker.ChainNode) (crypto.PubKey, error) {
+	keyJSONBytes, err := chainNode.ReadFile(ctx, "config/priv_validator_key.json")
+	if err != nil {
+		return nil, err
+	}
+	var pvKey cmprivval.FilePVKey
+	if err = cmtjson.Unmarshal(keyJSONBytes, &pvKey); err != nil {
+		return nil, fmt.Errorf("failed to unmarshal priv_validator_key.json: %w", err)
+	}
+	return pvKey.PubKey, nil
+}
+
+// fundBridgeNodeWallet funds the bridge node's DA wallet for blob submission
+func (suite *DockerIntegrationTestSuite) fundBridgeNodeWallet(ctx context.Context, bridgeNode types.DANode) {
+	// hack to get around global, need to set the address prefix before use.
+	sdk.GetConfig().SetBech32PrefixForAccount("celestia", "celestiapub")
+
+	suite.T().Log("Funding bridge node DA wallet...")
+	fundingWallet := suite.celestiaChain.GetFaucetWallet()
+
+	// Get the bridge node's wallet
+	bridgeWallet, err := bridgeNode.GetWallet()
+	suite.Require().NoError(err)
+
+	// fund the bridge node wallet
+	daFundingAmount := sdk.NewCoins(sdk.NewCoin("utia", math.NewInt(10_000_000)))
+	err = suite.sendFunds(ctx, suite.celestiaChain, fundingWallet, bridgeWallet, daFundingAmount, 0)
+	suite.Require().NoError(err)
+}
